Resolve the Decidim source path with @snyk/gemfile
Change the Decidim source path resolving to use @snyk/gemfile
instead of running `bundle` in a child process. Running `bundle`
won't work when ruby/bundler is not yet installed.

This allows resolving the Decidim source path from the
`Gemfile.lock` of the project when Decidim is installed either
from git or from a file path.